Warning for 3rd party WP plugins & shadow cookies
This adds a warning for 3rd party plugins that don't include the shadow cookies in WP REST requests.  I encountered this issue with the plugin Mailchimp for Woocommerce, it was calling the production instance because the shadow cookie was not set. The production instance did not have this plugin installed thus the REST request errored out. 

For me this was not a serious problem, but I can image scenarios where a plugin was installed in production and could be modified by actions from the staging instance of that plugin.
 
I resolved my issue by adding [a line](https://github.com/mailchimp/mc-woocommerce/blob/master/bootstrap.php#L951) with this code:
```
$headers['cookie'] = 'wpp_shadow=<MY_SHADOW_ID>; seravo_shadow=<MY_SHADOW_ID>;';
```

I think the general case should be documented here since it is a Seravo-specific shadow cookie issue that might occur in other plugins too.