Bugfix: Don't show trainings banner in all 3 languages
Fix the logic so that there is just one trainings banner, and it is either
shown or not. The old way of having 3 independent banners resulted in
showing all 3 banners for some users at the same time, which naturally
does not make sens. Now there is one simple set of controls for the banner
itself, and the text inside is automatically correct based on browser
settings.